requirements.txt
PyGithub==1.59.0
openai==1.10.0
python-dotenv==1.0.1
